join controller in one class and rename v2 dtos
