+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v2/FileMetadataV2Dto.java
@@ -12,7 +12,7 @@ import lombok.Data;
 import org.springframework.format.annotation.DateTimeFormat;
 
 @Data
-public class FileMetadataDto {
+public class FileMetadataV2Dto {
 
   @NotNull
   @NotBlank
