+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsfilereporter/controller/model/v2/FileReportV2Dto.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 import lombok.Data;
 
 @Data
-public class FileReportDto {
+public class FileReportV2Dto {
 
-  Collection<FileMetadataDto> filesRecentlyUploaded;
+  Collection<FileMetadataV2Dto> filesRecentlyUploaded;
 }
